+++ scripts/open-file-viewer-tab.sh
@@ -26,7 +26,7 @@ viewer_bin="$script_dir/../target/release/herdr-file-viewer"
 
 open_tab() {
   exec "$herdr_bin" plugin pane open \
-    --plugin herdr-file-viewer \
+    --plugin advanced-herdr-file-viewer \
     --entrypoint file-viewer \
     --placement tab \
     --focus